requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
python-pptx>=0.6.21
PyYAML>=6.0
